a2a_bedrock/requirements.txt
PyYAML
Jinja2
cfn-lint
openpyxl
boto3